Polish std/net release boundary
